backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
pip==21.3.1
pytz==2023.3.post1
setuptools==60.2.0
sqlparse==0.4.4
wheel==0.37.1